Avoid judge server crash caused by permission problems and give compile error for java files with package.

+++ judge/run/compile_java.sh
@@ -45,6 +45,10 @@ if [ -z "$MAINCLASS" ]; then
 	echo "Error: no 'main' found in any class file." >> log
 	exit 1
 fi
+if [ ! -f "$MAINCLASS.class" ]; then
+	echo "No $MAINCLASS.class generated." >> log
+	exit 1
+fi
 
 echo $MAINCLASS > mainclass.txt
 zip binary.zip mainclass.txt *.class
